+++ CollapseDisplay/AssemblyAttributes.cs
@@ -1,3 +1,3 @@
 ﻿using R2API.Utils;
 
-[assembly: NetworkCompatibility(CompatibilityLevel.NoNeedForSync, VersionStrictness.DifferentModVersionsAreOk)]
+[assembly: NetworkCompatibility(CompatibilityLevel.EveryoneMustHaveMod, VersionStrictness.EveryoneNeedSameModVersion)]