+++ CollapseDisplay/HealthBarType.cs
@@ -1,0 +1,11 @@
+﻿namespace CollapseDisplay
+{
+    public enum HealthBarType
+    {
+        Unknown,
+        Hud,
+        Combat,
+        AllyList,
+        Count
+    }
+}