+++ CollapseDisplay/HealthBarTypeProvider.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace CollapseDisplay
+{
+    public class HealthBarTypeProvider : MonoBehaviour
+    {
+        public HealthBarType Type;
+    }
+}